+++ src/main/java/org/broad/igv/session/SessionReader.java
@@ -35,7 +35,7 @@ import java.io.InputStream;
 public interface SessionReader {
 
     static boolean isSessionFile(String f) {
-        return f.endsWith(".xml") || f.endsWith(".php") || f.endsWith(".php3") || f.endsWith(".session");
+        return SessionMetadata.isSessionFile(f);
     }
 
     void loadSession(InputStream inputStream, Session session, String sessionPath) throws IOException;